# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Dora Explorer application.",
        "title": "Dora Explorer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/network/splits": {
            "get": {
                "description": "Returns information about active network forks/splits, their participation rates, and head blocks. On round-based networks, forks follow ready clients' reported heads; epoch_votes_unavailable is true and epoch vote/weight metrics are null. ready_client_count records supporting endpoints, not consensus stake.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "Get network splits",
                "operationId": "getNetworkSplits",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include stale forks with no voting weight (default: false)",
                        "name": "with_stale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APINetworkSplitsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/slot/{slotOrHash}/ptc_votes": {
            "get": {
                "description": "Returns the PTC (Payload Timeliness Committee) payload attestations included in a gloas+ block.\nThe attestations target the previous slot, mirroring what is shown on the slot page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slot"
                ],
                "summary": "Get PTC votes for a slot",
                "operationId": "getSlotPtcVotes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot number or block root (0x-prefixed hex)",
                        "name": "slotOrHash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resolve the PTC on the fork with this committee-shuffling dependent root (0x-prefixed hex)",
                        "name": "dependent_root",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Resolve the PTC on the fork that this block root sits on (0x-prefixed hex)",
                        "name": "block_root",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APISlotPtcVotesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid slot number or root format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.APINetworkSplitInfo": {
            "type": "object",
            "properties": {
                "epoch_votes_unavailable": {
                    "type": "boolean"
                },
                "fork_id": {
                    "type": "string"
                },
                "head_block_hash": {
                    "type": "string"
                },
                "head_execution_number": {
                    "type": "integer"
                },
                "head_root": {
                    "type": "string"
                },
                "head_slot": {
                    "type": "integer"
                },
                "is_canonical": {
                    "type": "boolean"
                },
                "last_epoch_participation": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "x-nullable": true
                },
                "last_epoch_votes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "x-nullable": true
                },
                "ready_client_count": {
                    "type": "integer"
                },
                "total_chain_weight": {
                    "type": "integer",
                    "x-nullable": true
                }
            }
        },
        "api.APINetworkSplitsData": {
            "type": "object",
            "properties": {
                "current_epoch": {
                    "type": "integer"
                },
                "current_slot": {
                    "type": "integer"
                },
                "finalized_epoch": {
                    "type": "integer"
                },
                "finalized_root": {
                    "type": "string"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APINetworkSplitInfo"
                    }
                }
            }
        },
        "api.APINetworkSplitsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.APINetworkSplitsData"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.APISlotPtcAggregate": {
            "type": "object",
            "properties": {
                "aggregation_bits": {
                    "type": "string"
                },
                "blob_data_available": {
                    "type": "boolean"
                },
                "payload_present": {
                    "type": "boolean"
                },
                "signature": {
                    "type": "string"
                },
                "unique_voter_count": {
                    "type": "integer"
                },
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APISlotPtcValidator"
                    }
                },
                "vote_count": {
                    "type": "integer"
                },
                "vote_percent": {
                    "type": "number"
                }
            }
        },
        "api.APISlotPtcValidator": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.APISlotPtcVotesData": {
            "type": "object",
            "properties": {
                "aggregates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APISlotPtcAggregate"
                    }
                },
                "block_root": {
                    "type": "string"
                },
                "duties_available": {
                    "type": "boolean"
                },
                "non_voter_count": {
                    "description": "Unvoted committee seats.",
                    "type": "integer"
                },
                "non_voter_percent": {
                    "type": "number"
                },
                "non_voters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APISlotPtcValidator"
                    }
                },
                "participation": {
                    "type": "number"
                },
                "slot": {
                    "type": "integer"
                },
                "total_ptc_size": {
                    "description": "Committee seats, including duplicates.",
                    "type": "integer"
                },
                "unique_non_voter_count": {
                    "type": "integer"
                },
                "unique_validator_count": {
                    "type": "integer"
                },
                "unique_voter_count": {
                    "type": "integer"
                },
                "vote_count": {
                    "description": "Total aggregate seat votes.",
                    "type": "integer"
                },
                "voted_block_root": {
                    "type": "string"
                },
                "voted_slot": {
                    "type": "integer"
                }
            }
        },
        "api.APISlotPtcVotesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.APISlotPtcVotesData"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}